requirements.txt
cassandra-driver==3.29.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
coverage==7.4.0
geomet==0.2.1.post1
idna==3.6
iniconfig==2.0.0
kafka-python==2.0.2
packaging==23.2
peewee==3.17.0
pluggy==1.3.0
py4j==0.10.9.7
pyspark==3.5.0
pytest==7.4.4
pytest-cov==2.12.1
rand==0.3.0
requests==2.31.0
six==1.16.0
toml==0.10.2
urllib3==2.1.0
